feat(database): refactor financial data types to INTEGER and update database schema

+++ lib/features/expenses/data/models/expense_model.dart
@@ -12,8 +12,9 @@ class ExpenseModel extends Expense {
   factory ExpenseModel.fromMap(Map<String, dynamic> map) {
     return ExpenseModel(
       id: map['id'] as int,
-      shiftId: map['shift_id'] as int, // تمت الإضافة
-      amount: (map['amount'] as num).toDouble(),
+      shiftId: map['shift_id'] as int,
+      // Refactored: تحويل إلى int 
+      amount: (map['amount'] as num).toInt(),
       reason: map['reason'] as String,
       createdAt: map['created_at'] as String,
     );
@@ -21,8 +22,8 @@ class ExpenseModel extends Expense {
 
   Map<String, dynamic> toMap() {
     return {
-      'shift_id': shiftId, // تمت الإضافة
-      'amount': amount,
+      'shift_id': shiftId,
+      'amount': amount, // سيتم حفظه كـ Integer
       'reason': reason,
       'created_at': createdAt,
     };
