+++ lib/features/expenses/domain/entities/expense.dart
@@ -2,8 +2,8 @@ import 'package:equatable/equatable.dart';
 
 class Expense extends Equatable {
   final int? id;
-  final int shiftId; // تمت الإضافة وتحديدها كـ مطلوب (Required)
-  final double amount;
+  final int shiftId;
+  final int amount; // Refactored: تغيير من double إلى int (Cents)
   final String reason;
   final String createdAt;
